Separate screens from components
