+++ src/screens/QuestionsPage/Question.jsx
@@ -1,6 +1,6 @@
 import { Link } from "react-router-dom";
 import styled from "styled-components";
-import { paper } from "../UI/css/paper";
+import { paper } from "../../components/UI/css/paper";
 
 const Question = (props) => {
   const date = new Date(props.question.date);
